requirements.txt
picozero
micropython-urequests
micropython-bme280
micropython-sgp40

